HS-97 REFACT : split count 함수 파일로 분리
