+++ src/init.c
@@ -24,7 +24,7 @@ void    init_minishell(void)
 			curr_token = make_token_list(splitted_input);
 			set_process_list(&g_minishell_info.ps_list, curr_token);
 			make_node_to_envp();
-			executor();
+			// executor();
 			free_all(replaced_input, splitted_input);
 		}
 	}
